Move classes to a package structure.
